Added Alert Queue for communication between Receiver and Acu
Replaced the old callback with a simple std::queue<IncomingAlert*> to
which the Receiver pushes Alerts. Acu::CheckForAlerts (name to be
discussed) checks these and processes them in Acu::OnReceive.
Further changes include moving the topic into IncomingAlert (performance
implications of this are still unclear) and changing tests to accomodate
for the new structure.